refactor(components): rename slotted-styles file to attachSlottedCss | bh
