+++ packages/components/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './a11y';
+export { attachSlottedCss } from './attachSlottedCss';
 export * from './attribute-observer';
 export { BREAKPOINTS, parseJSON } from './breakpoint-customizable';
 export { observeBreakpointChange, unobserveBreakpointChange } from './breakpoint-observer';
@@ -17,7 +18,6 @@ export { paramCaseToCamelCase } from './paramCaseToCamelCase';
 export * from './pds-fetch';
 export * from './property-observer';
 export * from './scrolling';
-export * from './slotted-styles';
 export * from './has-document';
 export * from './has-window';
 export * from './sync';
